+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 * The Singly logo is no longer used as the default loading indicator.
 
+* Fixed an issue where applying access tokens with the Singly API would always
+  create a new account.
+
 ## 1.2.1 - March 4th, 2013
 
 * Fixed example in documentation for the SinglyLoginPickerViewController.